feat(design): make @daffodil/design compatible with typescript@^4.0.0 (#1783)
