+++ libs/design/src/atoms/form/checkbox-set/checkbox-set.component.ts
@@ -17,7 +17,6 @@ import { DaffCheckboxComponent } from '../checkbox/checkbox.component';
   templateUrl: './checkbox-set.component.html',
   styleUrls: ['./checkbox-set.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-
 })
 export class DaffCheckboxSetComponent {
