+++ libs/design/src/atoms/form/form-label/form-label.directive.ts
@@ -6,9 +6,6 @@ import {
 @Directive({
   selector: '[daffFormLabel]',
 })
-
 export class DaffFormLabelDirective {
-
   @HostBinding('class.daff-form-label') class = true;
-
 }
