+++ libs/design/src/atoms/form/radio-set/radio-set.component.ts
@@ -10,7 +10,6 @@ import {
   templateUrl: './radio-set.component.html',
   styleUrls: ['./radio-set.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-
 })
 export class DaffRadioSetComponent {
