+++ libs/design/src/atoms/progress-indicator/progress-indicator.component.spec.ts
@@ -12,7 +12,7 @@ import {
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
-import { DaffPalette } from '../../core/colorable/colorable';
+import { DaffPalette } from '../../core/colorable/public_api';
 import { DaffProgressIndicatorComponent } from './progress-indicator.component';
 
 @Component({
