+++ libs/design/src/core/colorable/color-in-palette.ts
@@ -1,0 +1,5 @@
+import { DaffPaletteEnum } from './colorable-enum';
+
+export function colorInPalette(color: string): boolean{
+  return (<any>Object).values(DaffPaletteEnum).includes(color);
+}
