+++ libs/design/src/core/colorable/colorable-enum.ts
@@ -1,0 +1,10 @@
+export enum DaffPaletteEnum {
+  PRIMARY = 'primary',
+  SECONDARY = 'secondary',
+  ACCENT = 'accent', //TODO: damienwebdev Deprecate accent
+  TERTIARY = 'tertiary',
+  BLACK = 'black',
+  WHITE = 'white',
+  THEME = 'theme',
+  THEMECONTRAST = 'theme-contrast'
+}
