+++ libs/design/src/core/colorable/public_api.ts
@@ -1,0 +1,5 @@
+export {
+  DaffPalette,
+  DaffColorable,
+} from './colorable';
+export { daffColorMixin } from './colorable-mixin';
