+++ libs/design/src/molecules/card/card/card.component.ts
@@ -12,7 +12,7 @@ import {
   DaffColorable,
   DaffPalette,
   daffColorMixin,
-} from '../../../core/colorable/colorable';
+} from '../../../core/colorable/public_api';
 
 /**
  * An _elementRef and an instance of renderer2 are needed for the Colorable mixin
@@ -33,12 +33,14 @@ const _daffCardBase = daffColorMixin(DaffCardBase);
   templateUrl: './card.component.html',
   styleUrls: ['./card.component.scss'],
   encapsulation: ViewEncapsulation.None,
+  //todo(damienwebdev): remove once decorators hit stage 3 - https://github.com/microsoft/TypeScript/issues/7342
+  // eslint-disable-next-line @angular-eslint/no-inputs-metadata-property
+  inputs: ['color'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 
 export class DaffCardComponent extends _daffCardBase implements DaffColorable {
 
-  @Input() color: DaffPalette;
   @Input() raised = false;
 
   /**
